Fix type error from float to int

+++ unstrat/scripts/vpxd_mini.py
@@ -18,7 +18,7 @@ import pandas as pd
 from pathlib import Path
 
 # Passed arguments
-last = float(sys.argv[1])                     # only process last many outputs
+last = int(sys.argv[1])                        # only process last many outputs
 
 fig, axs = plt.subplots(2, 1, figsize=(6,8), dpi=300)
 runs = ['AB', 'BA']
